# inventory-api/docs/openapi/runtime-baseline.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "inventory-api runtime baseline",
    "version": "0.5.0",
    "description": "Baseline OpenAPI parcial y factual para superficies críticas verificadas en runtime. No cubre todo el monolito y se complementa con un manifiesto explícito de exclusiones contractuales."
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/receipts": {
      "get": {
        "operationId": "listPurchaseReceipts",
        "summary": "Listar documentos de recepción visibles para la compañía",
        "tags": [
          "receipts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-governance": {
          "routePolicy": "receipt.view",
          "actorScope": "company-admin"
        },
        "x-runtime-source": {
          "mountPath": "/api/receipts",
          "routeFile": "src/routes/receipt.routes.js",
          "expressPath": "/"
        },
        "responses": {
          "200": {
            "description": "Listado de recepciones"
          },
          "403": {
            "description": "Sin permisos"
          }
        }
      },
      "post": {
        "operationId": "createPurchaseReceipt",
        "summary": "Crear documento de recepción",
        "tags": [
          "receipts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-governance": {
          "routePolicy": "receipt.inspect",
          "actorScope": "company-admin"
        },
        "x-runtime-source": {
          "mountPath": "/api/receipts",
          "routeFile": "src/routes/receipt.routes.js",
          "expressPath": "/"
        },
        "responses": {
          "201": {
            "description": "Recepción creada"
          },
          "400": {
            "description": "Payload inválido"
          },
          "403": {
            "description": "Sin permisos"
          }
        }
      }
    },
    "/api/receipts/{id}": {
      "get": {
        "operationId": "getPurchaseReceipt",
        "summary": "Obtener detalle de documento de recepción",
        "tags": [
          "receipts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-governance": {
          "routePolicy": "receipt.view",
          "actorScope": "company-admin"
        },
        "x-runtime-source": {
          "mountPath": "/api/receipts",
          "routeFile": "src/routes/receipt.routes.js",
          "expressPath": "/:id"
        },
        "responses": {
          "200": {
            "description": "Detalle de recepción"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Recepción no encontrada"
          }
        }
      }
    },
    "/api/receipts/{id}/confirm": {
      "post": {
        "operationId": "confirmPurchaseReceipt",
        "summary": "Confirmar recepción e ingresar stock",
        "tags": [
          "receipts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-governance": {
          "routePolicy": "receipt.confirm",
          "actorScope": "company-admin"
        },
        "x-runtime-source": {
          "mountPath": "/api/receipts",
          "routeFile": "src/routes/receipt.routes.js",
          "expressPath": "/:id/confirm"
        },
        "responses": {
          "200": {
            "description": "Recepción confirmada"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Recepción no encontrada"
          },
          "409": {
            "description": "Estado inválido"
          }
        }
      }
    },
    "/api/receipts/{id}/reverse": {
      "post": {
        "operationId": "reversePurchaseReceipt",
        "summary": "Revertir recepción confirmada",
        "tags": [
          "receipts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-governance": {
          "routePolicy": "receipt.reverse",
          "actorScope": "company-admin"
        },
        "x-runtime-source": {
          "mountPath": "/api/receipts",
          "routeFile": "src/routes/receipt.routes.js",
          "expressPath": "/:id/reverse"
        },
        "responses": {
          "200": {
            "description": "Recepción revertida"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Recepción no encontrada"
          },
          "409": {
            "description": "Estado inválido"
          }
        }
      }
    }
  }
}